refactor: split appimage bundler in multiple files to support new backends
